Add Memory Node IDs to Merge/Split nodes using attributes (#14)
Attributes are a bit cumbersome due to the fact that they have to be
interned by the MLIR context. LLVM 18 did not yet have a finished
version of "properties" in MLIR, so that's where we are at.

There was also an inconsistency where some of the operations ended in
"Operation", while others did not. I